colors are fixed

+++ src/main/java/eu/ansquare/starr/util/math/ColorConversion.java
@@ -1,12 +1,14 @@
 package eu.ansquare.starr.util.math;
 
+import eu.ansquare.starr.StarR;
+
 import java.awt.*;
 
 public class ColorConversion {
-	public static float[] toScaledArray(Color color, int scale){
-		float red = (color.getRed() / 255) * scale;
-		float green = (color.getGreen() / 255) * scale;
-		float blue = (color.getBlue() / 255) * scale;
+	public static float[] toScaledArray(Color color, float scale){
+		float red = (color.getRed() / 255f) * scale;
+		float green = (color.getGreen() / 255f) * scale;
+		float blue = (color.getBlue() / 255f) * scale;
 		return new float[] {red, green, blue};
 	}
 }
